Cleanup

+++ Krauler/Logger.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Globalization;
 using System.IO;
-using OpenQA.Selenium.DevTools;
 
 namespace Krauler
 {
@@ -23,10 +22,7 @@ namespace Krauler
         public Logger(string logFile)
         {
             LogFile = logFile;
-            if (File.Exists(LogFile))
-            {
-                File.Delete(LogFile);
-            }
+            if (File.Exists(LogFile)) File.Delete(LogFile);
         }
 
         public string LogFile { get; set; }
